Remove : header 중복되는 파일 삭제
