+++ src/layouts/CommonLayout.tsx
@@ -1,12 +1,11 @@
 import Footer from '@/components/Footer/Footer';
 import Header from '@/components/Header/Header';
-import Header2 from '@/components/Header/Header2';
 import { Outlet } from 'react-router-dom';
 
 export default function CommonLayout() {
   return (
     <div>
-      <Header2 />
+      <Header />
       <main className='mainContent'>
         <Outlet />
       </main>
